fix reset links script
